% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mean.R
\name{compute_mean}
\alias{compute_mean}
\title{compute_mean}
\usage{
compute_mean(x)
}
\arguments{
\item{x}{input value}
}
\value{
the mean of x
}
\description{
Gets the mean of a vector
}
\examples{
compute_mean(3)
compute_mean(c(1,2,4,3))

}
