% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear_regress}
\alias{linear_regress}
\title{linear_regress}
\usage{
linear_regress(y, x)
}
\arguments{
\item{x, y}{input value}
}
\value{
regression result
}
\description{
compute linear regression models using least square method, and get a result which includes estimated coefficients,standard error of estimated coefficients, t statistics and p-value..
}
\examples{
library(MASS)  # make sure MASS is loaded
data(Boston)
y=Boston$medv
x <- as.matrix(Boston[, -which(names(Boston) == "medv")])
linear_regress(y,x)

}
