% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PredictY}
\alias{PredictY}
\title{PredictY}
\usage{
PredictY(x, beta_hat)
}
\arguments{
\item{x, beta_hat}{input value}
}
\value{
y_hat
}
\description{
Gets predict y for a linear regression
}
\examples{
mat <- matrix(c(1, 2, 3, 4, 5, 6,7,8,9), nrow = 3);
beta = c(1.1, 2.2, 3.3);
y_hat=PredictY(mat,beta);

}
