% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PredictY}
\alias{PredictY}
\title{PredictY}
\usage{
PredictY(x, beta_hat)
}
\arguments{
\item{x}{A numeric matrix or data frame representing the covariates (without any NA values).}

\item{beta_hat}{A numeric vector of estimated regression coefficients.}
}
\value{
PredictY produces a vector of predictions.
}
\description{
Gets predicted values for a linear regression.
}
\examples{
mat=matrix(c(1, 2, 3, 4, 5, 6,7,8,9), nrow = 3,byrow=TRUE);
beta = c(1.1, 2.2, 3.3,4.4);
PredictY(mat,beta);

#Using Boston dataset
#medv is dependent variable
#other variables in Boston are independent variables
require(MASS)
data(Boston)
y=Boston$medv
x=as.matrix(Boston[, -which(names(Boston) == "medv")])
lm1=linear_regress(y,x)
PredictY(x,lm1[["Estimate"]])

#Using mtcars dataset
#mpg is dependent variable
#other variables in mtcars are independent variables
data("mtcars")
y=mtcars$mpg
x=as.matrix(mtcars[, -which(names(mtcars) == "mpg")])
lm1=linear_regress(y,x)
PredictY(x,lm1[["Estimate"]])

#Using airquality dataset, and omit NA values
#Ozone is dependent variable
#Temp, Wind, Solar.R, Month are independent variables
data(airquality)
airquality_clean <- na.omit(airquality)
y=airquality_clean$Ozone
x=cbind(airquality_clean$Temp,airquality_clean$Wind,airquality_clean$Solar.R,airquality_clean$Month)
lm1=linear_regress(y,x)
PredictY(x,lm1[["Estimate"]])

}
